rxn 프리퍼런스 추가.
