+++ lib/common/data/preference/item/nullable_preference_item.dart
@@ -1,0 +1,5 @@
+import 'package:fast_app_base/common/data/preference/item/preference_item.dart';
+
+class NullablePreferenceItem<T> extends PreferenceItem<T?> {
+  NullablePreferenceItem(String key, [T? defaultValue]) : super(key, defaultValue);
+}
