+++ lib/common/data/preference/prefs.dart
@@ -1,5 +1,6 @@
-import 'package:fast_app_base/common/data/app_preferences.dart';
 import 'package:fast_app_base/common/theme/custom_theme.dart';
+
+import 'item/preference_item.dart';
 
 class Prefs {
   static final appTheme = PreferenceItem<CustomTheme>('appTheme', CustomTheme.light);
