+++ lib/common/theme/theme_util.dart
@@ -1,4 +1,4 @@
-import 'package:fast_app_base/common/data/prefs.dart';
+import 'package:fast_app_base/common/data/preference/prefs.dart';
 import 'package:fast_app_base/common/theme/custom_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
