+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
 import 'app.dart';
-import 'common/data/app_preferences.dart';
+import 'common/data/preference/app_preferences.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
